Mike | refactor the transformations to be a selector. MOved the column definitions and sort into a columnDefinitions folder as separate modules.
