+++ components/Table/columnDefinitions/index.js
@@ -1,0 +1,3 @@
+import playerMatchesColumns from './playerMatchesColumns';
+
+export { playerMatchesColumns };
