+++ components/Table/columnDefinitions/playerMatchesColumns.js
@@ -1,15 +1,5 @@
 import React from 'react';
-
-const defaultSort = (array, field) => array.sort((a, b) => {
-  if (a[field] < b[field]) return -1;
-  if (a[field] > b[field]) return 1;
-  return 0;
-});
-
-const useOriginalValueSort = (array, field) => {
-  const origField = field.slice(0, field.lastIndexOf('Display'));
-  return defaultSort(array, origField);
-};
+import { defaultSort, useOriginalValueSort } from './utility';
 
 export default [{
   displayName: 'ID',
