+++ selectors/index.js
@@ -1,4 +1,9 @@
 import getAlphabetizedLinks from './getAlphabetizedLinks';
 import sortPlayerMatches from './sortPlayerMatches';
+import transformPlayerMatches from './transformPlayerMatches';
 
-export { getAlphabetizedLinks, sortPlayerMatches };
+export {
+  getAlphabetizedLinks,
+  sortPlayerMatches,
+  transformPlayerMatches,
+};
